+++ packages/firebase_auth/example/lib/main.dart
@@ -41,10 +41,9 @@ class _MyHomePageState extends State<MyHomePage> {
   Future<String> _testSignInAnonymously() async {
     final FirebaseUser user = await _auth.signInAnonymously();
     assert(user != null);
-    assert(user == _auth.currentUser);
     assert(user.isAnonymous);
     assert(!user.isEmailVerified);
-    assert(await user.getToken() != null);
+    assert(await user.getIdToken() != null);
     if (Platform.isIOS) {
       // Anonymous auth doesn't show up as a provider on iOS
       assert(user.providerData.isEmpty);
@@ -57,6 +56,10 @@ class _MyHomePageState extends State<MyHomePage> {
       assert(user.providerData[0].photoUrl == null);
       assert(user.providerData[0].email == null);
     }
+
+    final FirebaseUser currentUser = await _auth.currentUser();
+    assert(user.uid == currentUser.uid);
+
     return 'signInAnonymously succeeded: $user';
   }
 
@@ -71,7 +74,11 @@ class _MyHomePageState extends State<MyHomePage> {
     assert(user.email != null);
     assert(user.displayName != null);
     assert(!user.isAnonymous);
-    assert(await user.getToken() != null);
+    assert(await user.getIdToken() != null);
+
+    final FirebaseUser currentUser = await _auth.currentUser();
+    assert(user.uid == currentUser.uid);
+
     return 'signInWithGoogle succeeded: $user';
   }