+++ packages/firebase_auth/lib/firebase_auth.dart
@@ -4,12 +4,13 @@
 
 import 'dart:async';
 
-import 'package:meta/meta.dart';
 import 'package:flutter/services.dart';
+import 'package:meta/meta.dart';
 
 /// Represents user data returned from an identity provider.
 class UserInfo {
   final Map<String, dynamic> _data;
+
   UserInfo._(this._data);
 
   /// The provider identifier.
@@ -36,6 +37,7 @@ class UserInfo {
 /// Represents a user.
 class FirebaseUser extends UserInfo {
   final List<UserInfo> providerData;
+
   FirebaseUser._(Map<String, dynamic> data)
       : providerData = data['providerData']
             .map((Map<String, dynamic> info) => new UserInfo._(info))
@@ -53,8 +55,8 @@ class FirebaseUser extends UserInfo {
   /// Obtains the id token for the current user, forcing a [refresh] if desired.
   ///
   /// Completes with an error if the user is signed out.
-  Future<String> getToken({bool refresh: false}) {
-    return FirebaseAuth.channel.invokeMethod('getToken', <String, bool>{
+  Future<String> getIdToken({bool refresh: false}) {
+    return FirebaseAuth.channel.invokeMethod('getIdToken', <String, bool>{
       'refresh': refresh,
     });
   }
@@ -71,12 +73,38 @@ class FirebaseAuth {
     'plugins.flutter.io/firebase_auth',
   );
 
+  final Map<int, StreamController<FirebaseUser>> _authStateChangedControllers =
+      <int, StreamController<FirebaseUser>>{};
+
   /// Provides an instance of this class corresponding to the default app.
   ///
   /// TODO(jackson): Support for non-default apps.
   static FirebaseAuth instance = new FirebaseAuth._();
 
-  FirebaseAuth._();
+  FirebaseAuth._() {
+    channel.setMethodCallHandler(_callHandler);
+  }
+
+  /// Receive [FirebaseUser] each time the user signIn or signOut
+  Stream<FirebaseUser> get onAuthStateChanged {
+    Future<int> _handle;
+
+    StreamController<FirebaseUser> controller;
+    controller = new StreamController<FirebaseUser>.broadcast(onListen: () {
+      _handle = channel.invokeMethod('startListeningAuthState');
+      _handle.then((int handle) {
+        _authStateChangedControllers[handle] = controller;
+      });
+    }, onCancel: () {
+      _handle.then((int handle) async {
+        await channel.invokeMethod(
+            "stopListeningAuthState", <String, int>{"id": handle});
+        _authStateChangedControllers.remove(handle);
+      });
+    });
+
+    return controller.stream;
+  }
 
   /// Asynchronously creates and becomes an anonymous user.
   ///
@@ -90,8 +118,8 @@ class FirebaseAuth {
   Future<FirebaseUser> signInAnonymously() async {
     final Map<String, dynamic> data =
         await channel.invokeMethod('signInAnonymously');
-    _currentUser = new FirebaseUser._(data);
-    return _currentUser;
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
   }
 
   Future<FirebaseUser> createUserWithEmailAndPassword({
@@ -107,8 +135,8 @@ class FirebaseAuth {
         'password': password,
       },
     );
-    _currentUser = new FirebaseUser._(data);
-    return _currentUser;
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
   }
 
   Future<FirebaseUser> signInWithEmailAndPassword({
@@ -124,8 +152,19 @@ class FirebaseAuth {
         'password': password,
       },
     );
-    _currentUser = new FirebaseUser._(data);
-    return _currentUser;
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  Future<FirebaseUser> signInWithFacebook(
+      {@required String accessToken}) async {
+    assert(accessToken != null);
+    final Map<String, dynamic> data =
+        await channel.invokeMethod('signInWithFacebook', <String, String>{
+      'accessToken': accessToken,
+    });
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
   }
 
   Future<FirebaseUser> signInWithGoogle({
@@ -141,17 +180,97 @@ class FirebaseAuth {
         'accessToken': accessToken,
       },
     );
-    _currentUser = new FirebaseUser._(data);
-    return _currentUser;
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  Future<FirebaseUser> signInWithCustomToken({@required String token}) async {
+    assert(token != null);
+    final Map<String, dynamic> data = await channel.invokeMethod(
+      'signInWithCustomToken',
+      <String, String>{
+        'token': token,
+      },
+    );
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
   }
 
   Future<Null> signOut() async {
-    await channel.invokeMethod("signOut");
-    _currentUser = null;
+    return await channel.invokeMethod("signOut");
   }
 
-  FirebaseUser _currentUser;
+  /// Asynchronously gets current user, or `null` if there is none.
+  Future<FirebaseUser> currentUser() async {
+    final Map<String, dynamic> data = await channel.invokeMethod("currentUser");
+    final FirebaseUser currentUser =
+        data == null ? null : new FirebaseUser._(data);
+    return currentUser;
+  }
 
-  /// Synchronously gets the cached current user, or `null` if there is none.
-  FirebaseUser get currentUser => _currentUser;
+  /// Links email account with current user and returns [Future<FirebaseUser>]
+  /// basically current user with addtional email infomation
+  ///
+  /// throws [PlatformException] when
+  /// 1. email address is already used
+  /// 2. wrong email and password provided
+  Future<FirebaseUser> linkWithEmailAndPassword({
+    @required String email,
+    @required String password,
+  }) async {
+    assert(email != null);
+    assert(password != null);
+    final Map<String, dynamic> data = await channel.invokeMethod(
+      'linkWithEmailAndPassword',
+      <String, String>{
+        'email': email,
+        'password': password,
+      },
+    );
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  /// Links google account with current user and returns [Future<FirebaseUser>]
+  ///
+  /// throws [PlatformException] when
+  /// 1. No current user provided (user has not logged in)
+  /// 2. No google credentials were found for given [idToken] and [accessToken]
+  /// 3. Google account already linked with another [FirebaseUser]
+  /// Detailed documentation on possible error causes can be found in [Android docs](https://firebase.google.com/docs/reference/android/com/google/firebase/auth/FirebaseUser#exceptions_4) and [iOS docs](https://firebase.google.com/docs/reference/ios/firebaseauth/api/reference/Classes/FIRUser#/c:objc(cs)FIRUser(im)linkWithCredential:completion:)
+  /// TODO: Throw custom exceptions with error codes indicating cause of exception
+  Future<FirebaseUser> linkWithGoogleCredential({
+    @required String idToken,
+    @required String accessToken,
+  }) async {
+    assert(idToken != null);
+    assert(accessToken != null);
+    final Map<String, dynamic> data = await channel.invokeMethod(
+      'linkWithGoogleCredential',
+      <String, String>{
+        'idToken': idToken,
+        'accessToken': accessToken,
+      },
+    );
+    final FirebaseUser currentUser = new FirebaseUser._(data);
+    return currentUser;
+  }
+
+  Future<Null> _callHandler(MethodCall call) async {
+    switch (call.method) {
+      case "onAuthStateChanged":
+        _onAuthStageChangedHandler(call);
+        break;
+    }
+    return null;
+  }
+
+  void _onAuthStageChangedHandler(MethodCall call) {
+    final Map<String, dynamic> data = call.arguments["user"];
+    final int id = call.arguments["id"];
+
+    final FirebaseUser currentUser =
+        data != null ? new FirebaseUser._(data) : null;
+    _authStateChangedControllers[id].add(currentUser);
+  }
 }