+++ packages/firebase_database/lib/src/event.dart
@@ -25,7 +25,7 @@ class Event {
 /// A DataSnapshot contains data from a Firebase Database location.
 /// Any time you read Firebase data, you receive the data as a DataSnapshot.
 class DataSnapshot {
-  Map<String, dynamic> _data;
+  final Map<String, dynamic> _data;
   DataSnapshot._(this._data);
 
   /// The key of the location that generated this DataSnapshot.
@@ -33,4 +33,34 @@ class DataSnapshot {
 
   /// Returns the contents of this data snapshot as native types.
   dynamic get value => _data['value'];
+}
+
+class MutableData {
+  final Map<String, dynamic> _data;
+  @visibleForTesting
+  MutableData.private(this._data);
+
+  /// The key of the location that generated this MutableData.
+  String get key => _data['key'];
+
+  /// Returns the mutable contents of this MutableData as native types.
+  dynamic get value => _data['value'];
+  set value(dynamic newValue) => _data['value'] = newValue;
+}
+
+/// A DatabaseError contains code, message and details of a Firebase Database
+/// Error that results from a transaction operation at a Firebase Database
+/// location.
+class DatabaseError {
+  Map<String, dynamic> _data;
+  DatabaseError._(this._data);
+
+  /// One of the defined status codes, depending on the error.
+  int get code => _data['code'];
+
+  /// A human-readable description of the error.
+  String get message => _data['message'];
+
+  /// Human-readable details on the error and additional information.
+  String get details => _data['details'];
 }