+++ packages/android_intent/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2
+
+* Add support for transferring structured Dart values into Android Intent
+  instances as extra Bundle data.
+
 ## 0.0.1
 
 * Initial release