+++ packages/device_info/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.0.3 - October 24, 2017
+
+* Add support for utsname
+
+## 0.0.2 - October 20, 2017
+
+* Fixed broke type comparison
+* Added "isPhysicaldevice" field, detecting emulators/simulators
+
 ## 0.0.1 - June 28, 2017
 
 * Implements platform-specific device/OS properties