+++ packages/firebase_auth/CHANGELOG.md
@@ -1,3 +1,44 @@
+## 0.3.1
+
+* Change GMS dependency to 11.+
+
+## 0.3.0
+
+* **Breaking Change**: Method FirebaseUser getToken was renamed to getIdToken.
+
+## 0.2.5
+
+* Added support for linkWithCredential with Google credential
+
+## 0.2.4
+
+* Added support for `signInWithCustomToken`
+* Added `Stream<FirebaseUser> onAuthStateChanged` event to listen when the user change
+
+## 0.2.3+1
+
+* Aligned author name with rest of repo.
+
+## 0.2.3
+
+* Remove dependency on Google/SignIn
+
+## 0.2.2
+
+* Remove dependency on FirebaseUI
+
+## 0.2.1
+
+* Added support for linkWithEmailAndPassword
+
+## 0.2.0
+
+* **Breaking Change**: Method currentUser is async now.
+
+## 0.1.2
+
+* Added support for signInWithFacebook
+
 ## 0.1.1
 
 * Updated to Firebase SDK to always use latest patch version for 11.0.x builds