+++ packages/local_auth/CHANGELOG.md
@@ -1,3 +1,7 @@
+## [0.0.2]
+
+* Support stickyAuth mode.
+
 ## [0.0.1] - 6/21/2017
 
 * Initial release of local authentication plugin.